reservation_count_create

+++ lesson/manage-lesson.php
@@ -52,30 +52,33 @@
                 <?php
 
                 $sql2 = "SELECT * FROM tbl_lesson where account_id ='$account_id'";
+                $sql3 = "SELECT COUNT(lesson_id) AS lesson_id FROM tbl_form GROUP BY lesson_id;";
 
                 $rec2 = mysqli_query($connect, $sql2);
+                $rec3 = mysqli_query($connect, $sql3);
 
-
-                if($rec2==TRUE)
+                if($rec2 && $rec3 ==TRUE)
                 {
                     $count = mysqli_num_rows($rec2); // Function to get all the rows in database
+                    $count2 = mysqli_num_rows($rec3);
 
                     $on=1;
 
+
                     if($count>0)
                     {
-                        while($rows=mysqli_fetch_assoc($rec2))
+                        while($rows=mysqli_fetch_array($rec2)) while($rows2=mysqli_fetch_array($rec3))
                         {
                             $course = $rows['course'];
                             $content = $rows['content'];
                             $day = $rows['day'];
-                            $form_id = $count;
+                            $lesson = $rows2['lesson_id'];
                             ?>
                             <tr>
                                 <td style="text-align: center"><?php echo $course; ?></td>
                                 <td style="text-align: center"><?php echo $content; ?></td>
                                 <td style="text-align: center"><?php echo $day; ?></td>
-                                <td style="text-align: center"><?php echo $form_id?></td>
+                                <td style="text-align: center"><?php echo $lesson?></td>
                                 <td>
                                     <a class="btn-primary" href="update-lesson.php?account_id=<?= $account_id=$_GET['account_id']?>&lesson_id=<?= $lesson_id?>"> Update your Lesson</a>
                                     <a class="btn-secondary" href="delete-lesson.php?account_id=<?= $account_id=$_GET['account_id']?>&lesson_id=<?= $lesson_id?>"> Delete your Lesson</a>
